move codet into separate header file
The codet class is currently shared between the front-end code classes and
the classes that represent goto-instructions.  This commit moves this class
into a separate header file so it can be consumed by both without including
everything else.
